requirements.txt
chumpy==0.70
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
datasets==2.8.0
diffusers==0.18.1
easydict==1.13
einops==0.6.1
imageio==2.31.1
kornia==0.6.8
matplotlib==3.7.2
mxnet==1.9.1
omegaconf==2.3.0
onnx==1.16.1
opencv-python==4.8.0.74
opencv-python-headless==4.5.5.64
pandas==1.5.3
Pillow==9.3.0
protobuf==3.20.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.8.1
tensorboard==2.11.2
timm==1.0.7
tokenizers==0.12.1
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchelastic==0.2.2
torchmetrics==1.0.0
torchtext==0.14.1
torchvision==0.13.0+cu113
transformers==4.30.2
